Upgrade to v1.22.0 of the SpotInst Terraform Provider

+++ sdk/dotnet/Aws/Inputs/ElastigroupIntegrationEcsAutoscaleDownGetArgs.cs
@@ -16,7 +16,7 @@ namespace Pulumi.SpotInst.Aws.Inputs
         public Input<int>? EvaluationPeriods { get; set; }
 
         [Input("maxScaleDownPercentage")]
-        public Input<int>? MaxScaleDownPercentage { get; set; }
+        public Input<double>? MaxScaleDownPercentage { get; set; }
 
         public ElastigroupIntegrationEcsAutoscaleDownGetArgs()
         {
